js: remove some dynamic change to function table
In the new llvm compiler of emscripten dynamically adding/removing
functions to the global table is very slow.  The recommended way is to
pre-define all the functions once.

There are still a few more to take care of.

The code is a bit ugly, I have to rely on a global js variable to store
the results of the functions.  I could probably use a closure instead
but I couldn't find a clean way to make it work.

Since we create the table at startup I also had to increase the static
table size.