Start to remove obj->id support
This is not useful for sky objects, since they have the most generic
designations methods.

At the moment the id is only used for module object and for landscape
and skycultures.

This commit keeps the id attribute but turns it into a const char * that
is not handled by the object.  It also remove the id argument to all the
object creation functions.

The goal is to eventually remove it totally.