Attempt at better UTC -> TT conversion
Specially at dates that are outside the leap second range.  I am trying
to maintain DUT1 in the [-1, +1] range.  The previous algorithm was
making DUT1 blow up when we set time in long past of future, resulting in
some rendering bugs.

There are a few other accuracy issues because our ΔT function is not
very good around 2020.