Use stencil hack to fix rendering of large mesh
We use the stencil buffer to make sure we never render overlapping
triangles of a single mesh.  Normal there should be no such triangles,
but they tend to show up when we use mesh subdivision.

Only enabled if the mesh has been subdivided, and only with the libtess2
tessellation for the moment.