Core navigation handles time speed
If we set the core.time_speed property we directly control the time
increment at each frame.  This makes it easier to integrate time
animation.

Also added core.y_offset attribute, that is not used for the moment, but
could be used to implement rendering offset.

Eventually all those attributes should be moved to a navigation module I
think.