Add project_to_win and project_to_clip functions
Those will replace the deprecated 'project' function, with a cleaner
API, more similar to gluProject.

The idea is that all the projections will first convert the view
coordinates into an intermediate representation that can then be
projected using a standard opengl perspective projection.  The big
advantage with this is that the depth value should then work just like a
normal OpenGL application.