Flipped projection uses the 4x4 matrix
Instead of having to test the flags.  That way the code is closer to an
normal OpenGL code where we would implement the flip directly in the
perspective projection.

Also changed the unprojection function to apply the invert perspective
projection first, since all the projections are now of the form:

    PerspectiveProj * NonLinearProj

The invert is better expressed as:

    Inv(NonLinearProj) * Inv(PerspectiveProj)

Might introduce regressions.