Mount orientation stored as mat instead of a quat
The problem with a quaternion is that it cannot properly represent an
axis flip.

If the mount is alt/az I would like the transformation to be the
identity (mount == observed) and if the mount is equatorial, I would
like the transformation to be the same as h2i (horizontal to ICRF).
Since h2i is left handed we cannot really represent it as a quaternion.

I removed the support for ecliptic frame for the moment, since we never
use it.  Also removed the support for slewing between mounts for the
moment.