Add support for comet data in stel jsonl format
So that we can extend it with additional data if needed.

We still support the MPC text data as well for now, so that we can
compare the performance and check for bugs.  Once the code is stable and
if there are no issue we can remove the support for the txt format.